refactor: move all signup protocols to a single file
