+++ src/presentation/controllers/signup/signup.spec.ts
@@ -1,8 +1,6 @@
 import SignUpController from './signup';
-import { InvalidParamError, MissingParamError, ServerError } from '../errors';
-import { EmailValidator } from '../protocols';
-import { AccountModel } from '../../domain/models/account';
-import { AddAccount, AddAccountModel } from '../../domain/usecases/add-account';
+import { InvalidParamError, MissingParamError, ServerError } from '../../errors';
+import { EmailValidator, AccountModel, AddAccount, AddAccountModel } from './signup-protocols';
 
 const makeEmailValidator = (): EmailValidator => {
   class EmailValidatorStub implements EmailValidator {
